Move routing path from server core to ktor-http
