+++ ktor-http/api/ktor-http.api
@@ -47,6 +47,7 @@ public final class io/ktor/http/CodecsKt {
 	public static final fun encodeURLParameter (Ljava/lang/String;Z)Ljava/lang/String;
 	public static synthetic fun encodeURLParameter$default (Ljava/lang/String;ZILjava/lang/Object;)Ljava/lang/String;
 	public static final fun encodeURLPath (Ljava/lang/String;)Ljava/lang/String;
+	public static final fun encodeURLPathComponent (Ljava/lang/String;)Ljava/lang/String;
 	public static final fun encodeURLQueryComponent (Ljava/lang/String;ZZLjava/nio/charset/Charset;)Ljava/lang/String;
 	public static synthetic fun encodeURLQueryComponent$default (Ljava/lang/String;ZZLjava/nio/charset/Charset;ILjava/lang/Object;)Ljava/lang/String;
 }
@@ -1293,5 +1294,41 @@ public final class io/ktor/http/content/WriterContent : io/ktor/http/content/Out
 
 public final class io/ktor/http/websocket/UtilsKt {
 	public static final fun websocketServerAccept (Ljava/lang/String;)Ljava/lang/String;
+}
+
+public final class io/ktor/routing/RoutingPath {
+	public static final field Companion Lio/ktor/routing/RoutingPath$Companion;
+	public synthetic fun <init> (Ljava/util/List;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public final fun getParts ()Ljava/util/List;
+	public fun toString ()Ljava/lang/String;
+}
+
+public final class io/ktor/routing/RoutingPath$Companion {
+	public final fun getRoot ()Lio/ktor/routing/RoutingPath;
+	public final fun parse (Ljava/lang/String;)Lio/ktor/routing/RoutingPath;
+}
+
+public final class io/ktor/routing/RoutingPathKt {
+	public static final fun parseRoutingParameterName (Ljava/lang/String;)Ljava/lang/String;
+}
+
+public final class io/ktor/routing/RoutingPathSegment {
+	public fun <init> (Ljava/lang/String;Lio/ktor/routing/RoutingPathSegmentKind;)V
+	public final fun component1 ()Ljava/lang/String;
+	public final fun component2 ()Lio/ktor/routing/RoutingPathSegmentKind;
+	public final fun copy (Ljava/lang/String;Lio/ktor/routing/RoutingPathSegmentKind;)Lio/ktor/routing/RoutingPathSegment;
+	public static synthetic fun copy$default (Lio/ktor/routing/RoutingPathSegment;Ljava/lang/String;Lio/ktor/routing/RoutingPathSegmentKind;ILjava/lang/Object;)Lio/ktor/routing/RoutingPathSegment;
+	public fun equals (Ljava/lang/Object;)Z
+	public final fun getKind ()Lio/ktor/routing/RoutingPathSegmentKind;
+	public final fun getValue ()Ljava/lang/String;
+	public fun hashCode ()I
+	public fun toString ()Ljava/lang/String;
+}
+
+public final class io/ktor/routing/RoutingPathSegmentKind : java/lang/Enum {
+	public static final field Constant Lio/ktor/routing/RoutingPathSegmentKind;
+	public static final field Parameter Lio/ktor/routing/RoutingPathSegmentKind;
+	public static fun valueOf (Ljava/lang/String;)Lio/ktor/routing/RoutingPathSegmentKind;
+	public static fun values ()[Lio/ktor/routing/RoutingPathSegmentKind;
 }
