+++ ktor-http/common/src/io/ktor/http/RoutingPath.kt
@@ -1,10 +1,10 @@
 /*
- * Copyright 2014-2019 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2014-2020 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package io.ktor.routing
+package io.ktor.http
 
-import io.ktor.http.*
+import io.ktor.util.*
 
 /**
  * Represents a parsed routing path. Consist of number of segments [parts]
@@ -60,3 +60,17 @@ enum class RoutingPathSegmentKind {
     Parameter
 }
 
+/**
+ * Parses a name out of segment specification
+ */
+@InternalAPI
+fun parseRoutingParameterName(value: String): String {
+    val prefix = value.substringBefore('{', "")
+    val suffix = value.substringAfterLast('}', "")
+    val signature = value.substring(prefix.length + 1, value.length - suffix.length - 1)
+    return when {
+        signature.endsWith("?") -> signature.dropLast(1)
+        signature.endsWith("...") -> signature.dropLast(3)
+        else -> signature
+    }
+}
